+++ src/core/entities/user/error.go
@@ -1,0 +1,9 @@
+package user
+
+import "errors"
+
+//Password incorrect
+var ErrIncorrectPassword = errors.New("password incorrect")
+
+//Not found user error
+var ErrUserNotFound = errors.New("user not found")
