+++ src/core/useCases/user/interface.go
@@ -1,15 +1,16 @@
-package user
+package user_use_case
 
 import (
-	entity "go-api/src/core/entities"
+	entity_root "go-api/src/core/entities"
+	entity "go-api/src/core/entities/user"
 )
 
 //Repository interface
 type Repository interface {
-	GetById(id entity.ID) (*entity.User, error)
+	GetById(id entity_root.ID) (*entity.User, error)
 }
 
 //UseCase interface
 type UseCase interface {
-	GetUser(id entity.ID) (*entity.User, error)
+	GetUser(id entity_root.ID) (*entity.User, error)
 }
