+++ src/infra/repositories/user/main.go
@@ -1,7 +1,8 @@
 package model
 
 import (
-	entity "go-api/src/core/entities"
+	entity_root "go-api/src/core/entities"
+	entity "go-api/src/core/entities/user"
 
 	"github.com/jinzhu/gorm"
 )
@@ -12,7 +13,7 @@ type (
 	}
 
 	RepositoryUser interface {
-		GetById(id entity.ID) (user *entity.User, er error)
+		GetById(id entity_root.ID) (user *entity.User, er error)
 	}
 )
 
@@ -21,7 +22,7 @@ func NewUserModel(db *gorm.DB) (repository RepositoryUser) {
 	return &repositoryUser{db}
 }
 
-func (repository *repositoryUser) GetById(id entity.ID) (user *entity.User, er error) {
+func (repository *repositoryUser) GetById(id entity_root.ID) (user *entity.User, er error) {
 	user = &entity.User{}
 	repository.db.First(user, "id = ?", id)
 	return
